src/provtoolutils/pyproject.toml
[build-system]
requires = ["setuptools==62.3.2"]
build-backend = "setuptools.build_meta"

[project]
name = "provtoolutils"
description = "Model and utilities for storing provenance with data"
version = "0.16.4"
authors = [
    { name = "Frank Dressel"}
]
dependencies = [
  "cryptography==39.0.0",
  "GitPython==3.1.27",
  "importlib-metadata==4.11.4",
  "jsonschema==4.6.0",
  "prov==2.0.0",
  "python-dateutil==2.8.2",
  "requests==2.28.2",
]
requires-python = ">=3.7"

[project.urls]
Homepage = "https://github.com/dlr-sp/provtool"

[project.optional-dependencies]
dev = [
  "tox",
]

[tool.setuptools]
packages = [ "provtoolutils" ]
